refactor: organise as `Server` and `ServerFileSystem` classes
