+++ tests/test_message.py
@@ -1,6 +1,7 @@
+import time
 from ipaddress import IPv4Address
 from uuid import uuid4
-import time
+
 from remote_file_system.message import (
     Message,
     ReadFileRequest,
